backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# AWS SDK
boto3==1.34.34
aioboto3==12.3.0

# OpenAI
openai==1.12.0

# Data processing
pandas==2.1.4
numpy==1.26.3
pyarrow==15.0.0

# HTTP client
httpx==0.26.0

# Authentication & OAuth
authlib==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
phonenumbers==8.13.27
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
pylint==3.0.3
mypy==1.8.0

